document new example
