+++ docs/src/webknossos-py/examples/annotation_project_administration.md
@@ -1,0 +1,9 @@
+# Annotation Project Administration
+
+This example uses the [`Project` class](../../api/webknossos/administration/project.md#Project) and [`Task` class](../../api/webknossos/administration/task.md#task) to check annotation task status and submit new tasks.
+
+```python
+--8<--
+webknossos/examples/annotation_project_administration.py
+--8<--
+```
